Arreglado un cliente puede acceder a vehículos de otros clientes

+++ src/main/java/org/springframework/samples/petclinic/web/VehiculoController.java
@@ -13,6 +13,8 @@ import org.springframework.samples.petclinic.model.TipoVehiculo;
 import org.springframework.samples.petclinic.model.Vehiculo;
 import org.springframework.samples.petclinic.service.ClienteService;
 import org.springframework.samples.petclinic.service.VehiculoService;
+import org.springframework.samples.petclinic.service.exceptions.DuplicatedMatriculaException;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.ModelMap;
 import org.springframework.validation.BindingResult;
@@ -42,27 +44,20 @@ public class VehiculoController {
 	@GetMapping(value = { "/listadoVehiculos" })
 	public String listadoVehiculos(ModelMap model) {
 		String vista = "vehiculos/listadoVehiculos";
-		Iterable<Vehiculo> vehiculos = vehiculoService.findAll();
-		model.put("vehiculos", vehiculos);
+		String username = SecurityContextHolder.getContext().getAuthentication().getName();
+		Optional<Cliente> cliente = clienteService.findClientesByUsername(username);
+		List<Vehiculo> vehiculos = null;
+		if(cliente.isPresent()) {
+			vehiculos = vehiculoService.findVehiculosCliente(cliente.get());
+			model.put("vehiculos", vehiculos);
+		} else {
+			vehiculos = vehiculoService.findAll();
+			model.put("vehiculos", vehiculos);
+		}
 		return vista;
 		
 	}
-	
-	@GetMapping(value = { "/listadoVehiculos/{username}" })
-	public String listadoVehiculosCliente(@PathVariable(value="username") String username, ModelMap model) {
-		String vista = "vehiculos/listadoVehiculos";
-		Optional<Cliente> cliente = clienteService.findClientesByUsername(username);
-		if(cliente.isPresent()) {
-			List<Vehiculo> vehiculos = vehiculoService.findVehiculosCliente(cliente.get());
-			model.put("vehiculos", vehiculos);
-			return vista;
-		}else {
-			return "";
-		}
-		
-		
-	}
-	
+
 	
 	@GetMapping(value = "/new")
 	public String crearVehiculo(ModelMap model) {
@@ -74,12 +69,32 @@ public class VehiculoController {
 	@PostMapping(value = "/save")
 	public String guardarVehiculo(@Valid Vehiculo vehiculo, BindingResult result, ModelMap model) {
 		String vista;
-		
+		String username = SecurityContextHolder.getContext().getAuthentication().getName();
+		Optional<Cliente> cliente = clienteService.findClientesByUsername(username);
 		if(result.hasErrors()) {
 			model.addAttribute("vehiculo", vehiculo);
 			vista = "vehiculos/editVehiculo";
 		} else {
-			vehiculoService.saveVehiculo(vehiculo);
+			
+			//preguntar esta parte, ahora mismo no deja añadir o editar
+			
+			if(cliente.isPresent()) {  
+				vehiculo.setCliente(cliente.get());
+			} else {
+				model.addAttribute("message", "Ha habido un error con el cliente");
+				vista = listadoVehiculos(model);
+				return vista;
+			} 
+			
+			//
+			
+			
+			try { //comprobar que la matrícula no está duplicada
+				vehiculoService.saveVehiculo(vehiculo);
+			} catch(DuplicatedMatriculaException ex) {
+				result.rejectValue("matricula", "Ya existe un vehículo con la misma matrícula", "Ya existe un vehículo con la misma matrícula");
+				return "vehiculos/editVehiculo";
+			}
 			model.addAttribute("message", "Vehiculo created successfully");
 			vista = listadoVehiculos(model);
 		}
@@ -93,8 +108,12 @@ public class VehiculoController {
 		String vista;
 		Optional<Vehiculo> op = vehiculoService.findVehiculoById(id);
 		if(op.isPresent()) {
-			vehiculoService.delete(op.get());
-			model.addAttribute("message", "Vehiculo deleted successfully");
+			if(!op.get().getCitas().isEmpty()) {
+				model.addAttribute("message", "No se puede borrar un vehículo con una cita asociada");
+			} else {
+				vehiculoService.delete(op.get());
+				model.addAttribute("message", "Vehiculo deleted successfully");
+			}
 		} else {
 			model.addAttribute("message", "Vehiculo not found");
 			
@@ -119,42 +138,6 @@ public class VehiculoController {
 		return vista;
 	}
 	
-	
-	@PostMapping(value = "/save/{username}")
-	public String guardarVehiculoParaCliente(@Valid Vehiculo vehiculo, BindingResult result, ModelMap model, 
-			@PathVariable("username") String username) {
-		String vista;
-		
-		if(result.hasErrors()) {
-			model.addAttribute("vehiculo", vehiculo);
-			vista = "vehiculos/editVehiculo";
-		} else {
-			vehiculo.setCliente(clienteService.findClientesByUsername(username).get());
-			vehiculoService.saveVehiculo(vehiculo);
-			model.addAttribute("message", "Vehiculo created successfully");
-			vista = listadoVehiculosCliente(username, model);
-		}
-		
-		return vista;
-	}
-	
-	
-	
-	@GetMapping(value = "/delete/{username}/{vehiculoId}")
-	public String borrarVehiculoParaCliente(@PathVariable("username") String username, @PathVariable("vehiculoId") int vehiculoId, 
-			ModelMap model) {
-		String vista;
-		Optional<Vehiculo> op = vehiculoService.findVehiculoById(vehiculoId);
-		if(op.isPresent()) {
-			vehiculoService.delete(op.get());
-			model.addAttribute("message", "Vehiculo deleted successfully");
-		} else {
-			model.addAttribute("message", "Vehiculo not found");
-			
-		}
-		vista = listadoVehiculosCliente(username, model);
-		return vista;
-	}
 	
 	
 }
